// src/app/model/role.ts
export enum Roles {
    // SGB = 'SGB',
    // DEO = 'DEO',
    // SEO = 'SEO',
    // PEO = 'PEO',
    // PEM = 'PEM',
    // PARENT = 'PARENT',
    // MONITOR = 'MONITOR',
    // OBSERVER = 'OBSERVER',
    // OFFICE_BEARERS = 'OFFICE_BEARERS',
    // PRINCIPAL = 'PRINCIPAL',
    // HO = 'HO'

    HO_DIRECTOR = 'HO DIRECTOR',
    HO_COORDINATOR = 'HO COORDINATOR',
    PROVINCIAL_MONITOR = 'PROVINCIAL MONITOR',
    CFO = 'CFO',
    PROCUREMENT_COORDINATOR = 'PROCUREMENT COORDINATOR',
    FINANCE = 'FINANCE',
    DISTRICT_COORDINATOR = 'DISTRICT COORDINATOR',
    DISTRICT_DIRECTOR = 'DISTRICT DIRECTOR',
    DISTRICT_CES = 'DISTRICT CES',
    DISTRICT_MONITOR = 'DISTRICT MONITOR',
    SCHOOL_PRINCIPAL = 'SCHOOL PRINCIPAL',
    SGB_CHAIRPERSON = 'SGB CHAIRPERSON',
    SCHOOL_COORDINATOR = 'SCHOOL COORDINATOR',
    FOOD_HANDLER = 'FOOD HANDLER',
    LEARNER = 'LEARNER',
    SUPPLIER = 'SUPPLIER'    

}